Conflicts:
	Projects/TestBed/Classes/Tests/PackManagerTest.cpp

Former-commit-id: df88a895259cea07761a146c95643d0d0ca4c29c

+++ Sources/Internal/Render/2D/GraphicFont.cpp
@@ -410,8 +410,8 @@ Font::StringMetrics GraphicFont::DrawStringToBuffer(const WideString& str,
     metrics.drawRect.dx += -metrics.drawRect.x + 1;
     metrics.drawRect.dy += -metrics.drawRect.y + 1;
 
-    metrics.height = int32(ceilf(lastY));
-    metrics.width = int32(ceilf(lastX));
+    metrics.height = int32(std::ceil(lastY));
+    metrics.width = int32(std::ceil(lastX));
     metrics.baseline = yOffset + int32(fontInternal->baselineHeight);
     return metrics;
 }
